backend/requirements.txt
asgiref==3.2.6
Django==2.1.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
